+++ bfvmm/include/intrinsics/intrinsics_x64.h
@@ -97,12 +97,6 @@ void __write_ldtr(uint16_t val);
 
 uint64_t __read_rsp(void);
 
-struct idt_t
-{
-    uint64_t limit : 16;
-    uint64_t base  : 64;
-};
-
 void __read_gdt(void *gdt);
 void __write_gdt(void *gdt);
 
@@ -290,144 +284,146 @@ public:
 // =============================================================================
 
 // Selector Fields
-#define SELECTOR_TI_FLAG                                          (0x0004)
-#define SELECTOR_RPL_FLAG                                         (0x0003)
-#define SELECTOR_INDEX                                            (0xFFF8)
-#define SELECTOR_UNUSABLE                                         (1 << 16)
+#define SELECTOR_TI_FLAG                                            (0x0004)
+#define SELECTOR_RPL_FLAG                                           (0x0003)
+#define SELECTOR_INDEX                                              (0xFFF8)
+#define SELECTOR_UNUSABLE                                           (1 << 16)
 
 // Segment Access Rights
-#define SEGMENT_ACCESS_RIGHTS_TYPE                                (0x000F)
-#define SEGMENT_ACCESS_RIGHTS_TYPE_TSS_BUSY                       (0x0002)
-#define SEGMENT_ACCESS_RIGHTS_TYPE_RWA                            (0x0003)
-#define SEGMENT_ACCESS_RIGHTS_TYPE_REA                            (0x000B)
-#define SEGMENT_ACCESS_RIGHTS_TYPE_TSS_AVAILABLE                  (0x0009)
-#define SEGMENT_ACCESS_RIGHTS_CODE_DATA_DESCRIPTOR                (0x0010)
-#define SEGMENT_ACCESS_RIGHTS_SYSTEM_DESCRIPTOR                   (0x0010)
-#define SEGMENT_ACCESS_RIGHTS_DPL                                 (0x0060)
-#define SEGMENT_ACCESS_RIGHTS_PRESENT                             (0x0080)
-#define SEGMENT_ACCESS_RIGHTS_RESERVED                            (0x0F00)
-#define SEGMENT_ACCESS_RIGHTS_L                                   (0x2000)
-#define SEGMENT_ACCESS_RIGHTS_DB                                  (0x4000)
-#define SEGMENT_ACCESS_RIGHTS_GRANULARITY                         (0x8000)
-#define SEGMENT_ACCESS_RIGHTS_GRANULARITY_PAGES                   (0x8000)
+#define SEGMENT_ACCESS_RIGHTS_TYPE                                  (0x000F)
+#define SEGMENT_ACCESS_RIGHTS_TYPE_TSS_BUSY                         (0x0002)
+#define SEGMENT_ACCESS_RIGHTS_TYPE_RW                               (0x0002)
+#define SEGMENT_ACCESS_RIGHTS_TYPE_RWA                              (0x0003)
+#define SEGMENT_ACCESS_RIGHTS_TYPE_RE                               (0x000A)
+#define SEGMENT_ACCESS_RIGHTS_TYPE_REA                              (0x000B)
+#define SEGMENT_ACCESS_RIGHTS_TYPE_TSS_AVAILABLE                    (0x0009)
+#define SEGMENT_ACCESS_RIGHTS_CODE_DATA_DESCRIPTOR                  (0x0010)
+#define SEGMENT_ACCESS_RIGHTS_SYSTEM_DESCRIPTOR                     (0x0010)
+#define SEGMENT_ACCESS_RIGHTS_DPL                                   (0x0060)
+#define SEGMENT_ACCESS_RIGHTS_PRESENT                               (0x0080)
+#define SEGMENT_ACCESS_RIGHTS_RESERVED                              (0x0F00)
+#define SEGMENT_ACCESS_RIGHTS_L                                     (0x2000)
+#define SEGMENT_ACCESS_RIGHTS_DB                                    (0x4000)
+#define SEGMENT_ACCESS_RIGHTS_GRANULARITY                           (0x8000)
+#define SEGMENT_ACCESS_RIGHTS_GRANULARITY_PAGES                     (0x8000)
 
 // RFLAGS
 // 64-ia-32-architectures-software-developer-manual, section 3.4.3
-#define RFLAGS_CF_CARRY_FLAG                                      (1 << 0)
-#define RFLAGS_PF_PARITY_FLAG                                     (1 << 2)
-#define RFLAGS_AF_AUXILIARY_CARRY_FLAG                            (1 << 4)
-#define RFLAGS_ZF_ZERO_FLAG                                       (1 << 6)
-#define RFLAGS_SF_SIGN_FLAG                                       (1 << 7)
-#define RFLAGS_TF_TRAP_FLAG                                       (1 << 8)
-#define RFLAGS_IF_INTERRUPT_ENABLE_FLAG                           (1 << 9)
-#define RFLAGS_DF_DIRECTION_FLAG                                  (1 << 10)
-#define RFLAGS_OF_OVERFLOW_FLAG                                   (1 << 11)
-#define RFLAGS_IOPL_PRIVILEGE_LEVEL                               (3 << 12)
-#define RFLAGS_NT_NESTED_TASK                                     (1 << 14)
-#define RFLAGS_RF_RESUME_FLAG                                     (1 << 16)
-#define RFLAGS_VM_VIRTUAL_8086_MODE                               (1 << 17)
-#define RFLAGS_AC_ALIGNMENT_CHECK_ACCESS_CONTROL                  (1 << 18)
-#define RFLAGS_VIF_VIRTUAL_INTERUPT_FLAG                          (1 << 19)
-#define RFLAGS_VIP_VIRTUAL_INTERUPT_PENDING                       (1 << 20)
-#define RFLAGS_ID_ID_FLAG                                         (1 << 21)
+#define RFLAGS_CF_CARRY_FLAG                                        (1ULL << 0)
+#define RFLAGS_PF_PARITY_FLAG                                       (1ULL << 2)
+#define RFLAGS_AF_AUXILIARY_CARRY_FLAG                              (1ULL << 4)
+#define RFLAGS_ZF_ZERO_FLAG                                         (1ULL << 6)
+#define RFLAGS_SF_SIGN_FLAG                                         (1ULL << 7)
+#define RFLAGS_TF_TRAP_FLAG                                         (1ULL << 8)
+#define RFLAGS_IF_INTERRUPT_ENABLE_FLAG                             (1ULL << 9)
+#define RFLAGS_DF_DIRECTION_FLAG                                    (1ULL << 10)
+#define RFLAGS_OF_OVERFLOW_FLAG                                     (1ULL << 11)
+#define RFLAGS_IOPL_PRIVILEGE_LEVEL                                 (3ULL << 12)
+#define RFLAGS_NT_NESTED_TASK                                       (1ULL << 14)
+#define RFLAGS_RF_RESUME_FLAG                                       (1ULL << 16)
+#define RFLAGS_VM_VIRTUAL_8086_MODE                                 (1ULL << 17)
+#define RFLAGS_AC_ALIGNMENT_CHECK_ACCESS_CONTROL                    (1ULL << 18)
+#define RFLAGS_VIF_VIRTUAL_INTERUPT_FLAG                            (1ULL << 19)
+#define RFLAGS_VIP_VIRTUAL_INTERUPT_PENDING                         (1ULL << 20)
+#define RFLAGS_ID_ID_FLAG                                           (1ULL << 21)
 
 // CR0
 // 64-ia-32-architectures-software-developer-manual, section 2.5
-#define CRO_PE_PROTECTION_ENABLE                                  (1 << 0)
-#define CR0_MP_MONITOR_COPROCESSOR                                (1 << 1)
-#define CR0_EM_EMULATION                                          (1 << 2)
-#define CR0_TS_TASK_SWITCHED                                      (1 << 3)
-#define CR0_ET_EXTENSION_TYPE                                     (1 << 4)
-#define CR0_NE_NUMERIC_ERROR                                      (1 << 5)
-#define CR0_WP_WRITE_PROTECT                                      (1 << 16)
-#define CR0_AM_ALIGNMENT_MASK                                     (1 << 18)
-#define CR0_NW_NOT_WRITE_THROUGH                                  (1 << 29)
-#define CR0_CD_CACHE_DISABLE                                      (1 << 30)
-#define CR0_PG_PAGING                                             (1 << 31)
+#define CRO_PE_PROTECTION_ENABLE                                    (1ULL << 0)
+#define CR0_MP_MONITOR_COPROCESSOR                                  (1ULL << 1)
+#define CR0_EM_EMULATION                                            (1ULL << 2)
+#define CR0_TS_TASK_SWITCHED                                        (1ULL << 3)
+#define CR0_ET_EXTENSION_TYPE                                       (1ULL << 4)
+#define CR0_NE_NUMERIC_ERROR                                        (1ULL << 5)
+#define CR0_WP_WRITE_PROTECT                                        (1ULL << 16)
+#define CR0_AM_ALIGNMENT_MASK                                       (1ULL << 18)
+#define CR0_NW_NOT_WRITE_THROUGH                                    (1ULL << 29)
+#define CR0_CD_CACHE_DISABLE                                        (1ULL << 30)
+#define CR0_PG_PAGING                                               (1ULL << 31)
 
 // CR4
 // 64-ia-32-architectures-software-developer-manual, section 2.5
-#define CR4_VME_VIRTUAL8086_MODE_EXTENSIONS                       (1 << 0)
-#define CR4_PVI_PROTECTED_MODE_VIRTUAL_INTERRUPTS                 (1 << 1)
-#define CR4_TSD_TIME_STAMP_DISABLE                                (1 << 2)
-#define CR4_DE_DEBUGGING_EXTENSIONS                               (1 << 3)
-#define CR4_PSE_PAGE_SIZE_EXTENSIONS                              (1 << 4)
-#define CR4_PAE_PHYSICAL_ADDRESS_EXTENSIONS                       (1 << 5)
-#define CR4_MACHINE_CHECK_ENABLE                                  (1 << 6)
-#define CR4_PGE_PAGE_GLOBAL_ENABLE                                (1 << 7)
-#define CR4_PCE_PERFORMANCE_MONITOR_COUNTER_ENABLE                (1 << 8)
-#define CR4_OSFXSR                                                (1 << 9)
-#define CR4_OSXMMEXCPT                                            (1 << 10)
-#define CR4_VMXE_VMX_ENABLE_BIT                                   (1 << 13)
-#define CR4_SMXE_SMX_ENABLE_BIT                                   (1 << 14)
-#define CR4_FSGSBASE_ENABLE_BIT                                   (1 << 16)
-#define CR4_PCIDE_PCID_ENABLE_BIT                                 (1 << 17)
-#define CR4_OSXSAVE                                               (1 << 18)
-#define CR4_SMEP_SMEP_ENABLE_BIT                                  (1 << 20)
-#define CR4_SMAP_SMAP_ENABLE_BIT                                  (1 << 21)
-#define CR4_PKE_PROTECTION_KEY_ENABLE_BIT                         (1 << 22)
+#define CR4_VME_VIRTUAL8086_MODE_EXTENSIONS                         (1ULL << 0)
+#define CR4_PVI_PROTECTED_MODE_VIRTUAL_INTERRUPTS                   (1ULL << 1)
+#define CR4_TSD_TIME_STAMP_DISABLE                                  (1ULL << 2)
+#define CR4_DE_DEBUGGING_EXTENSIONS                                 (1ULL << 3)
+#define CR4_PSE_PAGE_SIZE_EXTENSIONS                                (1ULL << 4)
+#define CR4_PAE_PHYSICAL_ADDRESS_EXTENSIONS                         (1ULL << 5)
+#define CR4_MACHINE_CHECK_ENABLE                                    (1ULL << 6)
+#define CR4_PGE_PAGE_GLOBAL_ENABLE                                  (1ULL << 7)
+#define CR4_PCE_PERFORMANCE_MONITOR_COUNTER_ENABLE                  (1ULL << 8)
+#define CR4_OSFXSR                                                  (1ULL << 9)
+#define CR4_OSXMMEXCPT                                              (1ULL << 10)
+#define CR4_VMXE_VMX_ENABLE_BIT                                     (1ULL << 13)
+#define CR4_SMXE_SMX_ENABLE_BIT                                     (1ULL << 14)
+#define CR4_FSGSBASE_ENABLE_BIT                                     (1ULL << 16)
+#define CR4_PCIDE_PCID_ENABLE_BIT                                   (1ULL << 17)
+#define CR4_OSXSAVE                                                 (1ULL << 18)
+#define CR4_SMEP_SMEP_ENABLE_BIT                                    (1ULL << 20)
+#define CR4_SMAP_SMAP_ENABLE_BIT                                    (1ULL << 21)
+#define CR4_PKE_PROTECTION_KEY_ENABLE_BIT                           (1ULL << 22)
 
 // 64-ia-32-architectures-software-developer-manual, section 35.1
 // IA-32 Architectural MSRs
-#define IA32_DEBUGCTL_MSR                                         0x000001D9
-#define IA32_SYSENTER_CS_MSR                                      0x00000174
-#define IA32_SYSENTER_ESP_MSR                                     0x00000175
-#define IA32_SYSENTER_EIP_MSR                                     0x00000176
-#define IA32_PAT_MSR                                              0x00000277
-#define IA32_EFER_MSR                                             0xC0000080
-#define IA32_FS_BASE_MSR                                          0xC0000100
-#define IA32_GS_BASE_MSR                                          0xC0000101
+#define IA32_DEBUGCTL_MSR                                           0x000001D9
+#define IA32_SYSENTER_CS_MSR                                        0x00000174
+#define IA32_SYSENTER_ESP_MSR                                       0x00000175
+#define IA32_SYSENTER_EIP_MSR                                       0x00000176
+#define IA32_PAT_MSR                                                0x00000277
+#define IA32_EFER_MSR                                               0xC0000080
+#define IA32_FS_BASE_MSR                                            0xC0000100
+#define IA32_GS_BASE_MSR                                            0xC0000101
 
 // 64-ia-32-architectures-software-developer-manual, section 6.3.1
 // IA-32 Interrupts and Exceptions
-#define INTERRUPT_DIVIDE_ERROR                                        (0)
-#define INTERRUPT_DEBUG_EXCEPTION                                     (1)
-#define INTERRUPT_NMI_INTERRUPT                                       (2)
-#define INTERRUPT_BREAKPOINT                                          (3)
-#define INTERRUPT_OVERFLOW                                            (4)
-#define INTERRUPT_BOUND_RANGE_EXCEEDED                                (5)
-#define INTERRUPT_INVALID_OPCODE                                      (6)
-#define INTERRUPT_DEVICE_NOT_AVAILABLE                                (7)
-#define INTERRUPT_DOUBLE_FAULT                                        (8)
-#define INTERRUPT_COPROCESSOR_SEGMENT_OVERRUN                         (9)
-#define INTERRUPT_INVALID_TSS                                         (10)
-#define INTERRUPT_SEGMENT_NOT_PRESENT                                 (11)
-#define INTERRUPT_STACK_SEGMENT_FAULT                                 (12)
-#define INTERRUPT_GENERAL_PROTECTION                                  (13)
-#define INTERRUPT_PAGE_FAULT                                          (14)
-#define INTERRUPT_FLOATING_POINT_ERROR                                (16)
-#define INTERRUPT_ALIGNMENT_CHECK                                     (17)
-#define INTERRUPT_MACHINE_CHECK                                       (18)
-#define INTERRUPT_SIMD_FLOATING_POINT_EXCEPTION                       (19)
-#define INTERRUPT_VIRTUALIZATION_EXCEPTION                            (20)
+#define INTERRUPT_DIVIDE_ERROR                                      (0)
+#define INTERRUPT_DEBUG_EXCEPTION                                   (1)
+#define INTERRUPT_NMI_INTERRUPT                                     (2)
+#define INTERRUPT_BREAKPOINT                                        (3)
+#define INTERRUPT_OVERFLOW                                          (4)
+#define INTERRUPT_BOUND_RANGE_EXCEEDED                              (5)
+#define INTERRUPT_INVALID_OPCODE                                    (6)
+#define INTERRUPT_DEVICE_NOT_AVAILABLE                              (7)
+#define INTERRUPT_DOUBLE_FAULT                                      (8)
+#define INTERRUPT_COPROCESSOR_SEGMENT_OVERRUN                       (9)
+#define INTERRUPT_INVALID_TSS                                       (10)
+#define INTERRUPT_SEGMENT_NOT_PRESENT                               (11)
+#define INTERRUPT_STACK_SEGMENT_FAULT                               (12)
+#define INTERRUPT_GENERAL_PROTECTION                                (13)
+#define INTERRUPT_PAGE_FAULT                                        (14)
+#define INTERRUPT_FLOATING_POINT_ERROR                              (16)
+#define INTERRUPT_ALIGNMENT_CHECK                                   (17)
+#define INTERRUPT_MACHINE_CHECK                                     (18)
+#define INTERRUPT_SIMD_FLOATING_POINT_EXCEPTION                     (19)
+#define INTERRUPT_VIRTUALIZATION_EXCEPTION                          (20)
 
 // Debug Control
 // 64-ia-32-architectures-software-developer-manual, section 35.1
-#define IA32_DEBUGCTL_LBR                                             (1 << 0)
-#define IA32_DEBUGCTL_BTF                                             (1 << 1)
-#define IA32_DEBUGCTL_TR                                              (1 << 6)
-#define IA32_DEBUGCTL_BTS                                             (1 << 7)
-#define IA32_DEBUGCTL_BTINT                                           (1 << 8)
-#define IA32_DEBUGCTL_BTS_OFF_OS                                      (1 << 9)
-#define IA32_DEBUGCTL_BTS_OFF_USER                                    (1 << 10)
-#define IA32_DEBUGCTL_FREEZE_LBRS_ON_PMI                              (1 << 11)
-#define IA32_DEBUGCTL_FREEZE_PERFMON_ON_PMI                           (1 << 12)
-#define IA32_DEBUGCTL_ENABLE_UNCORE_PMI                               (1 << 13)
-#define IA32_DEBUGCTL_FREEZE_WHILE_SMM                                (1 << 14)
-#define IA32_DEBUGCTL_RTM_DEBUG                                       (1 << 15)
+#define IA32_DEBUGCTL_LBR                                           (1ULL << 0)
+#define IA32_DEBUGCTL_BTF                                           (1ULL << 1)
+#define IA32_DEBUGCTL_TR                                            (1ULL << 6)
+#define IA32_DEBUGCTL_BTS                                           (1ULL << 7)
+#define IA32_DEBUGCTL_BTINT                                         (1ULL << 8)
+#define IA32_DEBUGCTL_BTS_OFF_OS                                    (1ULL << 9)
+#define IA32_DEBUGCTL_BTS_OFF_USER                                  (1ULL << 10)
+#define IA32_DEBUGCTL_FREEZE_LBRS_ON_PMI                            (1ULL << 11)
+#define IA32_DEBUGCTL_FREEZE_PERFMON_ON_PMI                         (1ULL << 12)
+#define IA32_DEBUGCTL_ENABLE_UNCORE_PMI                             (1ULL << 13)
+#define IA32_DEBUGCTL_FREEZE_WHILE_SMM                              (1ULL << 14)
+#define IA32_DEBUGCTL_RTM_DEBUG                                     (1ULL << 15)
 
 // EFER
 // 64-ia-32-architectures-software-developer-manual, section 35.1
-#define IA32_EFER_SCE                                                 (1 << 0)
-#define IA32_EFER_LME                                                 (1 << 8)
-#define IA32_EFER_LMA                                                 (1 << 10)
-#define IA32_EFER_NXE                                                 (1 << 11)
+#define IA32_EFER_SCE                                               (1ULL << 0)
+#define IA32_EFER_LME                                               (1ULL << 8)
+#define IA32_EFER_LMA                                               (1ULL << 10)
+#define IA32_EFER_NXE                                               (1ULL << 11)
 
 // Serial COM Port Addresses
 // http://wiki.osdev.org/Serial_Ports
-#define COM1_PORT                                                     0x3f8
-#define COM2_PORT                                                     0x2f8
-#define COM3_PORT                                                     0x3e8
-#define COM4_PORT                                                     0x2e8
+#define COM1_PORT                                                   0x3f8
+#define COM2_PORT                                                   0x2f8
+#define COM3_PORT                                                   0x3e8
+#define COM4_PORT                                                   0x2e8
 
 #endif