+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - New GDT class that provides a cleaner abstraction of the GDT
+- New IDT class that provides a cleaner abstraction of the IDT
 - New VMCS state classes that organize the creation of the VMM's state, as
   well as the Host VM's state. Future versions will also add a Guest VM
   state as well
@@ -13,6 +14,11 @@
 - Added TSS structure to intrinsics code. This structure is used by the
   custom GDT for the VMM, but is not actually used.
 - The VMM now uses it's own GDT instead of the GDT provided by the Host OS.
+- The VMM now uses it's own IDT instead of the IDT provided by the Host OS.
+- The VMM now uses it's own CR0 instead of the CR0 provided by the Host OS.
+- The VMM now uses it's own CR4 instead of the CR4 provided by the Host OS.
+- The VMM now uses it's own RFLAGS instead of the RFLAGS provided by the Host OS.
+- The VMM now uses it's own EFER MSR instead of the EFER MSR provided by the Host OS.
 
 ### Changed
 - The VMCS state classes are now shared by pointer (i.e. shared_ptr)
@@ -22,10 +28,15 @@
 - If a VM-entry failured occured, the exit handler would incorrectly read
   the error as unknown because it was not filtering the VM-entry failure
   bit
+- Some of the macros in the intrinsics file were causing unsigned intergers
+  because they hit touched bit 31. The macros have been expanded to 64bits
+  to prevent this
 
 ### Removed
 - The old GDT logic that was in the intrinsics_x64 has been removed. Please
   use the new GDT class as it has the same functionality, but more.
+- The old IDT logic that was in the intrinsics_x64 has been removed. Please
+  use the new IDT class as it has the same functionality, but more.
 - The old vmcs_intel_x64_state class has been removed in favor of inheritance
   to better support different state types (VMM, Host VM and Guest VM). Please
   use the subclasses instead, or inherit manually